page: add abandon-page-change-evt
Spawning a thread to poll for page change events is somewhat
problematic if there's no mechanism to stop that thread on timeout or
if the event ends up being unused. This change defers polling until
the event is selected for synchronization and adds an
abandon-page-change-event procedure to allow users to abandon the
polling should synchronization time out.